\name{baronum}
\alias{baronum}
\docType{data}
\title{Baromètre du Numérique (2023)}

\description{
Le Baromètre du Numérique est une étude de référence sur l’adoption par les Français des équipements et des usages numériques. Il est le fruit d’une collaboration entre l'Autorité de régulation des communications électroniques, des postes et de la distribution de la presse (Arcep), le Conseil général de l'économie (CGE), le Programme Société Numérique de l’ANCT, à laquelle s’est jointe l’Arcom en 2022.
}

\usage{data("baronum")}

\format{
  Data frame avec 4184 observations et 255 variables.
}

\source{
L'ensemble des fichiers et de la documentation est consultable à l'adresse suivante :
\url{https://www.data.gouv.fr/fr/datasets/barometre-du-numerique/}
}

\references{
Les publications de l'ARCEP à partir des différentes éditions des baromètres du numérique sont accessibles depuis l'adresse suivante :
\url{https://www.arcep.fr/cartes-et-donnees/nos-publications-chiffrees/barometre-du-numerique/le-barometre-du-numerique-edition-2023.html}
}

\keyword{datasets}
