\name{drees}
\alias{drees}
\docType{data}
\title{Baromètre d'opinion de la DREES (2022)}

\description{
Le Baromètre d'opinion de la DREES est une enquête annuelle réalisée depuis 2000 selon la méthode des quotas auprès d'un échantillon d'au moins 3000 personnes interrogées en face à face, représentatif de la population vivant en France métropolitaine. L'enquête permet de suivre l'opinion à l'égard de la santé, de la protection sociale et des inégalités sociales.
}

\usage{data("drees")}

\format{
  Data frame avec 82332 observations et 438 variables.
}

\details{
Le fichier compile les baromètres annuels de 2000 à 2022.
}

\source{
L'ensemble des fichiers et de la documentation (questionnaires, dictionnaire des variables, etc.) est consultable à l'adresse suivante :
\url{https://data.drees.solidarites-sante.gouv.fr/explore/dataset/431_le-barometre-d-opinion/information/}
}

\keyword{datasets}
