\name{epcv}
\alias{epcv}
\docType{data}
\title{
Enquête EPCV - Participation Culturelle et Sportive (Insee, 2003)
}

\description{
Enquête permanente des conditions de vie (EPCV), réalisée par l'INSEE en 2003, sur le thème de la participation culturelle et sportive.
}

\usage{data("EPCV_2003")}

\format{
  Data frame avec 5626 observations et 578 variables
}

\source{
L'ensemble des fichiers et de la documentation (questionnaires, dictionnaire des variables, etc.) est consultable à l'adresse suivante :
\url{https://www.insee.fr/fr/statistiques/2569722}
}

\keyword{datasets}
