\name{fecond}
\alias{fecond}
\docType{data}
\title{
Enquête "Fécondité - Contraception - Dysfonctions sexuelles" en France Métropolitaine (2013)}

\description{
Le dispositif de recherche FECOND analyse les évolutions des pratiques contraceptives et abortives en France tout en étendant les problématiques de recherche à d'autres questions de santé sexuelle et reproductive (infertilité, dysfonctions sexuelles, IST).
Un dispositif en quatre temps a été conçu pour répondre à cette problématique : un premier volet par téléphone, un deuxième sur Internet fixe, un troisième randomisé entre téléphone et Intenet fixe et un quatrième mené auprès du panel ELIPSS. C'est ce dernier volet dont il est question ici.
}

\usage{data("Fecond_2013")}

\format{
  Data frame avec 883 observations et 51 variables.
}

\details{
Ce jeu de données est un extrait des données intégrales de l'enquête, réalisé à des fins pédagogiques par le Centre de Données Socio-Politiques.

L'échantillon correspond à celui de l'enquête originelle, mais seule une partie des variables est fournie.
}

\source{
L'ensemble des fichiers et de la documentation (questionnaires, dictionnaire des variables, etc.) est consultable à l'adresse suivante :
\url{https://data.sciencespo.fr/dataset.xhtml?persistentId=doi:10.21410/7E4/PLP8HD}

L'échantillon correspond à celui de l'enquête originelle, mais seule une partie des variables est fournie.
}

\references{
Legleye Stéphane, Bajos Nathalie, Moreau Caroline, 2022, "Base de données pédagogique : Fécondité - Contraception - Dysfonctions sexuelles en France Métropolitaine (2013)", \url{https://doi.org/10.21410/7E4/PLP8HD}, data.sciencespo, V3, UNF:6:6B98K6BcdlKkdaBueMFkrA== [fileUNF]}

\keyword{datasets}
