\name{hdv}
\alias{hdv}
\docType{data}
\title{Enquête Histoire de vie (Insee, 2003)}

\description{
Enquête "Histoire de vie - Construction des identités", réalisée par l'Insee en 2003.
}

\usage{data("HDV_2003")}

\format{
  Data frame avec 8403 observations et 814 variables.
}

\source{
L'ensemble des fichiers et de la documentation (questionnaires, dictionnaire des variables, etc.) est consultable à l'adresse suivante :
\url{https://www.insee.fr/fr/statistiques/2532244}
}

\references{
Guérin-Pace France, Samuel Olivia, Ville Isabelle (eds.), 2009, \emph{En quête d'appartenances. L’enquête Histoire de vie sur la construction des identités}. Paris: Ined Éditions, 2009. \url{https://doi.org/10.4000/books.ined.1013}.
}

\keyword{datasets}
