\name{injep}
\alias{injep}
\docType{data}
\title{Baromètre national des pratiques sportives (2018)}

\description{
Le baromètre national des pratiques sportives est réalisé par l'INJEP et destiné à suivre les grandes évolutions de la pratique sportive des Français de 15 ans et plus.
}

\usage{data("injep")}

\format{
  Data frame avec 4061 observations et 36 variables.
}

\details{
Ce jeu de données est un extrait des données intégrales de l'enquête, réalisé à des fins pédagogiques par Progedo.

L'échantillon correspond à celui de l'enquête originelle, mais seule une partie des variables est fournie.

Les informations sur l'enquête sont disponibles ici : \url{https://data.progedo.fr/studies/doi/10.13144/lil-1496}
}

\source{
\url{https://www.progedo.fr/formation/ressources/}
}

\references{
Patricia Croutte, Jörg Müller, sous la direction de Sandra Hoibian (Crédoc), 2019, "Baromètre national des pratiques sportives 2018", \emph{Rapport d'études de l'INJEP}, n°2019/01. \url{https://injep.fr/publication/barometre-national-des-pratiques-sportives-2018/}
}

\keyword{datasets}
