\name{pcs2003e}
\alias{pcs2003e}
\docType{data}
\title{
Nomenclature des PCS 2003 (Liste Emboîtée)
}

\description{
Tableau des équivalences entre les différents niveaux de la nomenclature des professions et catégories socioprofessionnelles de l'INSEE (version 2003).
}

\usage{data("pcs2003e")}
\format{
  A data frame with 497 observations on the following 4 variables.
  \describe{
    \item{\code{pcs497}}{a character vector. Code de la catégorie socioprofessionnelle au niveau 4}
    \item{\code{pcs42}}{a character vector. Code de la catégorie socioprofessionnelle au niveau 3}
    \item{\code{pcs24}}{a character vector. Code de la catégorie socioprofessionnelle au niveau 2}
    \item{\code{pcs8}}{a character vector. Code de la catégorie socioprofessionnelle au niveau 1}
  }
}

\source{
Les informations sur la nomenclature 2003 des PCS de l'INSEE est accessible à l'adresse suivante:
\url{https://www.insee.fr/fr/information/2497952}
}

\keyword{datasets}
