\name{pcs2003l}
\alias{pcs2003l}
\docType{data}
\title{
Nomenclature des PCS 2003 (Liste Exhaustive)
}

\description{
Liste exhaustive des professions et catégories socioprofessionnelles (PCS) de la nomenclature 2003 de l'INSEE, aux niveaux 1, 2, 3 et 4, avec leur code et leur libellé.
}

\usage{data("pcs2003l")}
\format{
  A data frame with 571 observations on the following 3 variables.
  \describe{
    \item{\code{niveau}}{a character vector. Niveau (1, 2, 3 ou 4) de la nomenclature}
    \item{\code{code}}{a character vector. Code de la catégorie socioprofessionnelle}
    \item{\code{libelle}}{a character vector. Libellé de la catégorie socioprofessionnelle}
  }
}

\source{
Les informations sur la nomenclature 2003 des PCS de l'INSEE est accessible à l'adresse suivante:
\url{https://www.insee.fr/fr/information/2497952}
}

\keyword{datasets}
