\name{sven}
\alias{sven}
\docType{data}
\title{Enquête "Styles de vie et Environnement" (2017)}

\description{
L'enquête "Styles de vie et Environnement"" (SVEN) propose d’explorer conjointement les représentations et attitudes des Français à l’égard des questions d’environnement et l’impact environnemental de leurs modes de vie.
Elle s'est déroulée en novembre 2017, pendant la phase élargie du panel ELIPSS.
}

\usage{data("Sven_2017")}

\format{
  Data frame avec 2362 observations et 56 variables.
}

\details{
Ce jeu de données est un extrait des données intégrales de l'enquête, réalisé à des fins pédagogiques par le Centre de Données Socio-Politiques.

L'échantillon correspond à celui de l'enquête originelle, mais seule une partie des variables est fournie.
}

\source{
L'ensemble des fichiers et de la documentation (questionnaires, dictionnaire des variables, etc.) est consultable à l'adresse suivante :
\url{https://data.sciencespo.fr/dataset.xhtml?persistentId=doi:10.21410/7E4/OFVDJL}
}

\keyword{datasets}
