\name{tdt}
\alias{tdt}
\docType{data}
\title{Enquête "Le temps donné aux tableaux" (Passeron & Pedler, 1987)}

\description{
Données de l'enquête emmenée par Jean-Claude Passeron et Emmanuel Pedler en 1987. Les enquêteurs ont suivi les visiteurs du musée Granet d'Aix-en-Provence, chronomètre en main, avec pour objectif de mettre en évidence les ressorts sociaux du « temps donné aux tableaux ».
}

\usage{data("Tableaux")}

\format{
  Data frame avec 144 observations et 37 variables.
}

\source{
Les données sont accessibles depuis les annexes du la réédition numérique de l'ouvrage tiré de l'enquête. On y trouve également le dictionnaire des variables.

\url{https://books.openedition.org/enseditions/16817}
}

\references{
Passeron Jean-Claude, Pedler Emmanuel Pedler, 2019, \emph{Le temps donné aux tableaux}, édité par Samuel Coavoux, Igor Martinache, et Pierre Mercklé. Lyon: ENS Éditions, 2019. \url{https://doi.org/10.4000/books.enseditions.16682}.}

\keyword{datasets}
